Pass PHPStan level 5
Fixes argument type mismatches, including two real bugs:
- Model_WPOptions::get() called get_network_option() without the network
  id, so every multisite option lookup used the key as the network id
- reset/revoke grace period views passed true as json_encode() flags

Also: cast string/float values before long2ip(), chr(), str_repeat(),
set_transient() and esc_attr(); pass arrays rather than query strings to
WP_User_Query; pass 0 instead of null to the WP_User constructor;
convert the 'yes'/'no' autoload value to bool for add/update_option().

+++ classes/controller/totp.php
@@ -2,7 +2,8 @@
 
 namespace TFAuthLS;
 
-class Controller_TOTP {
+class Controller_TOTP
+{
 
 	const TIME_WINDOW_LENGTH = 30;
 
@@ -11,9 +12,10 @@ class Controller_TOTP {
 	 *
 	 * @return Controller_TOTP
 	 */
-	public static function shared() {
+	public static function shared()
+	{
 		static $_shared = null;
-		if ( $_shared === null ) {
+		if ($_shared === null) {
 			$_shared = new Controller_TOTP();
 		}
 		return $_shared;
@@ -29,8 +31,9 @@ class Controller_TOTP {
 	 * @param string[] $recovery An array of recovery codes as hex strings.
 	 * @param bool|int $vtime The timestamp of the verification code or false to use the current timestamp.
 	 */
-	public function activate_2fa( $user, $secret, $recovery, $vtime = false ): void {
-		if ( $vtime === false ) {
+	public function activate_2fa($user, $secret, $recovery, $vtime = false): void
+	{
+		if ($vtime === false) {
 			$vtime = Controller_Time::time();
 		}
 
@@ -40,12 +43,12 @@ class Controller_TOTP {
 			$wpdb->prepare(
 				"INSERT INTO `{$table}` (`user_id`, `secret`, `recovery`, `ctime`, `vtime`, `mode`) VALUES (%d, %s, %s, UNIX_TIMESTAMP(), %d, 'authenticator')",
 				$user->ID,
-				Model_Compat::hex2bin( $secret ),
+				Model_Compat::hex2bin($secret),
 				implode(
 					'',
 					array_map(
-						function ( $r ) {
-							return Model_Compat::hex2bin( $r );
+						function ($r) {
+							return Model_Compat::hex2bin($r);
 						},
 						$recovery
 					)
@@ -61,7 +64,7 @@ class Controller_TOTP {
 		 *
 		 * @param \WP_User $user The user.
 		 */
-		do_action( 'TFA_LS_2fa_activated', $user );
+		do_action('TFA_LS_2fa_activated', $user);
 	}
 
 	/**
@@ -72,43 +75,44 @@ class Controller_TOTP {
 	 * @param string   $code
 	 * @return bool|null Returns null if the user does not have 2FA enabled, false if the code is invalid, and true if valid.
 	 */
-	public function validate_2fa( $user, $code, $update = true ): ?bool {
+	public function validate_2fa($user, $code, $update = true): ?bool
+	{
 		global $wpdb;
 		$table  = Controller_DB::shared()->secrets;
-		$record = $wpdb->get_row( $wpdb->prepare( "SELECT * FROM `{$table}` WHERE `user_id` = %d FOR UPDATE", $user->ID ), ARRAY_A );
-		if ( ! $record ) {
+		$record = $wpdb->get_row($wpdb->prepare("SELECT * FROM `{$table}` WHERE `user_id` = %d FOR UPDATE", $user->ID), ARRAY_A);
+		if (! $record) {
 			return null;
 		}
 
-		if ( preg_match( '/^(?:[a-f0-9]{4}\s*){4}$/i', $code ) ) {
+		if (preg_match('/^(?:[a-f0-9]{4}\s*){4}$/i', $code)) {
 			// Recovery code
-			$code          = strtolower( preg_replace( '/\s/i', '', $code ) );
-			$recoveryCodes = str_split( strtolower( bin2hex( $record['recovery'] ) ), 16 );
-			$index         = array_search( $code, $recoveryCodes );
-			if ( $index !== false ) {
-				if ( $update ) {
-					unset( $recoveryCodes[ $index ] );
-					$updatedRecoveryCodes = implode( '', $recoveryCodes );
-					$wpdb->query( $wpdb->prepare( "UPDATE `{$table}` SET `recovery` = X%s WHERE `id` = %d", $updatedRecoveryCodes, $record['id'] ) );
+			$code          = strtolower(preg_replace('/\s/i', '', $code));
+			$recoveryCodes = str_split(strtolower(bin2hex($record['recovery'])), 16);
+			$index         = array_search($code, $recoveryCodes);
+			if ($index !== false) {
+				if ($update) {
+					unset($recoveryCodes[$index]);
+					$updatedRecoveryCodes = implode('', $recoveryCodes);
+					$wpdb->query($wpdb->prepare("UPDATE `{$table}` SET `recovery` = X%s WHERE `id` = %d", $updatedRecoveryCodes, $record['id']));
 				}
-				$wpdb->query( 'COMMIT' );
+				$wpdb->query('COMMIT');
 				return true;
 			}
-		} elseif ( preg_match( '/^(?:[0-9]{3}\s*){2}$/i', $code ) ) {
+		} elseif (preg_match('/^(?:[0-9]{3}\s*){2}$/i', $code)) {
 			// TOTP code
-			$code    = preg_replace( '/\s/i', '', $code );
-			$secret  = bin2hex( $record['secret'] );
-			$matches = $this->check_code( $secret, $code, floor( $record['vtime'] / self::TIME_WINDOW_LENGTH ) );
-			if ( $matches !== false ) {
-				if ( $update ) {
-					$wpdb->query( $wpdb->prepare( "UPDATE `{$table}` SET `vtime` = %d WHERE `id` = %d", $matches, $record['id'] ) );
+			$code    = preg_replace('/\s/i', '', $code);
+			$secret  = bin2hex($record['secret']);
+			$matches = $this->check_code($secret, $code, (int) floor($record['vtime'] / self::TIME_WINDOW_LENGTH));
+			if ($matches !== false) {
+				if ($update) {
+					$wpdb->query($wpdb->prepare("UPDATE `{$table}` SET `vtime` = %d WHERE `id` = %d", $matches, $record['id']));
 				}
-				$wpdb->query( 'COMMIT' );
+				$wpdb->query('COMMIT');
 				return true;
 			}
 		}
 
-		$wpdb->query( 'ROLLBACK' );
+		$wpdb->query('ROLLBACK');
 		return false;
 	}
 
@@ -122,30 +126,31 @@ class Controller_TOTP {
 	 * @param null|array $windows An array of time windows or null to use the default.
 	 * @return int|float|false The time window if matches, otherwise false.
 	 */
-	public function check_code( $secret, $code, $previous = null, $windows = null ): int|float|false {
-		$timeCode = floor( Controller_Time::time() / self::TIME_WINDOW_LENGTH );
+	public function check_code($secret, $code, $previous = null, $windows = null): int|float|false
+	{
+		$timeCode = floor(Controller_Time::time() / self::TIME_WINDOW_LENGTH);
 
-		if ( $windows === null ) {
+		if ($windows === null) {
 			$windows    = array();
-			$validRange = array( -1, 1 ); // 90 second range for authenticator
+			$validRange = array(-1, 1); // 90 second range for authenticator
 
 			$lowRange  = $validRange[0];
 			$highRange = $validRange[1];
-			for ( $i = 0; $i >= $lowRange; $i-- ) {
+			for ($i = 0; $i >= $lowRange; $i--) {
 				$windows[] = $timeCode + $i;
 			}
-			for ( $i = 1; $i <= $highRange; $i++ ) {
+			for ($i = 1; $i <= $highRange; $i++) {
 				$windows[] = $timeCode + $i;
 			}
 		}
 
-		foreach ( $windows as $w ) {
-			if ( $previous !== null && $previous >= $w ) {
+		foreach ($windows as $w) {
+			if ($previous !== null && $previous >= $w) {
 				continue;
 			}
 
-			$expectedCode = $this->_generate_totp( $secret, dechex( $w ) );
-			if ( hash_equals( $expectedCode, $code ) ) {
+			$expectedCode = $this->_generate_totp($secret, dechex($w));
+			if (hash_equals($expectedCode, $code)) {
 				return $w * self::TIME_WINDOW_LENGTH;
 			}
 		}
@@ -161,19 +166,20 @@ class Controller_TOTP {
 	 * @param int $digits The number of digits.
 	 * @return string The TOTP value.
 	 */
-	private function _generate_totp( $key, string $time, $digits = 6 ): string {
-		$time = Model_Compat::hex2bin( str_pad( $time, 16, '0', STR_PAD_LEFT ) );
-		$key  = Model_Compat::hex2bin( $key );
-		$hash = hash_hmac( 'sha1', $time, $key );
+	private function _generate_totp($key, string $time, $digits = 6): string
+	{
+		$time = Model_Compat::hex2bin(str_pad($time, 16, '0', STR_PAD_LEFT));
+		$key  = Model_Compat::hex2bin($key);
+		$hash = hash_hmac('sha1', $time, $key);
 
-		$offset       = hexdec( substr( $hash, -2 ) ) & 0xf;
-		$intermediate = ( ( ( hexdec( substr( $hash, $offset * 2, 2 ) ) & 0x7f ) << 24 ) |
-			( ( hexdec( substr( $hash, ( $offset + 1 ) * 2, 2 ) ) & 0xff ) << 16 ) |
-			( ( hexdec( substr( $hash, ( $offset + 2 ) * 2, 2 ) ) & 0xff ) << 8 ) |
-			( ( hexdec( substr( $hash, ( $offset + 3 ) * 2, 2 ) ) & 0xff ) )
+		$offset       = hexdec(substr($hash, -2)) & 0xf;
+		$intermediate = (((hexdec(substr($hash, $offset * 2, 2)) & 0x7f) << 24) |
+			((hexdec(substr($hash, ($offset + 1) * 2, 2)) & 0xff) << 16) |
+			((hexdec(substr($hash, ($offset + 2) * 2, 2)) & 0xff) << 8) |
+			((hexdec(substr($hash, ($offset + 3) * 2, 2)) & 0xff))
 		);
-		$otp          = $intermediate % pow( 10, $digits );
+		$otp          = $intermediate % pow(10, $digits);
 
-		return str_pad( "{$otp}", $digits, '0', STR_PAD_LEFT );
+		return str_pad("{$otp}", $digits, '0', STR_PAD_LEFT);
 	}
 }
