+++ classes/model/crypto/base2n.php
@@ -25,7 +25,8 @@ namespace TFAuthLS\Crypto;
  *
  * @package binary-to-text-php
  */
-class Model_Base2n {
+class Model_Base2n
+{
 
 	protected $_chars;
 	protected $_bitsPerCharacter;
@@ -57,37 +58,37 @@ class Model_Base2n {
 		$padCharacter = '='
 	) {
 		// Ensure validity of $chars
-		if ( ! is_string( $chars ) || ( $charLength = strlen( $chars ) ) < 2 ) {
-			throw new \InvalidArgumentException( '$chars must be a string of at least two characters' );
+		if (! is_string($chars) || ($charLength = strlen($chars)) < 2) {
+			throw new \InvalidArgumentException('$chars must be a string of at least two characters');
 		}
 
 		// Ensure validity of $padCharacter
-		if ( $padFinalGroup ) {
-			if ( ! is_string( $padCharacter ) || ! isset( $padCharacter[0] ) ) {
-				throw new \InvalidArgumentException( '$padCharacter must be a string of one character' );
+		if ($padFinalGroup) {
+			if (! is_string($padCharacter) || ! isset($padCharacter[0])) {
+				throw new \InvalidArgumentException('$padCharacter must be a string of one character');
 			}
 
-			$padCharFound = $caseSensitive ? strpos( $chars, $padCharacter[0] ) : stripos( $chars, $padCharacter[0] );
+			$padCharFound = $caseSensitive ? strpos($chars, $padCharacter[0]) : stripos($chars, $padCharacter[0]);
 
-			if ( $padCharFound !== false ) {
-				throw new \InvalidArgumentException( '$padCharacter can not be a member of $chars' );
+			if ($padCharFound !== false) {
+				throw new \InvalidArgumentException('$padCharacter can not be a member of $chars');
 			}
 		}
 
 		// Ensure validity of $bitsPerCharacter
-		if ( ! is_int( $bitsPerCharacter ) ) {
-			throw new \InvalidArgumentException( '$bitsPerCharacter must be an integer' );
+		if (! is_int($bitsPerCharacter)) {
+			throw new \InvalidArgumentException('$bitsPerCharacter must be an integer');
 		}
-		if ( $bitsPerCharacter < 1 ) {
+		if ($bitsPerCharacter < 1) {
 			// $bitsPerCharacter must be at least 1
-			throw new \InvalidArgumentException( '$bitsPerCharacter can not be less than 1' );
+			throw new \InvalidArgumentException('$bitsPerCharacter can not be less than 1');
 		}
-		if ( $charLength < 1 << $bitsPerCharacter ) {
+		if ($charLength < 1 << $bitsPerCharacter) {
 			// Character length of $chars is too small for $bitsPerCharacter
 			// Find greatest acceptable value of $bitsPerCharacter
 			$bitsPerCharacter = 1;
 			$radix            = 2;
-			while ( $charLength >= ( $radix <<= 1 ) && $bitsPerCharacter < 8 ) {
+			while ($charLength >= ($radix <<= 1) && $bitsPerCharacter < 8) {
 				++$bitsPerCharacter;
 			}
 			$radix >>= 1;
@@ -98,9 +99,9 @@ class Model_Base2n {
 			);
 		}
 
-		if ( $bitsPerCharacter > 8 ) {
+		if ($bitsPerCharacter > 8) {
 			// $bitsPerCharacter must not be greater than 8
-			throw new \InvalidArgumentException( '$bitsPerCharacter can not be greater than 8' );
+			throw new \InvalidArgumentException('$bitsPerCharacter can not be greater than 8');
 		} else {
 			$radix = 1 << $bitsPerCharacter;
 		}
@@ -119,13 +120,14 @@ class Model_Base2n {
 	 *
 	 * @param   string $rawString  Binary data to encode
 	 */
-	public function encode( $rawString ): string {
+	public function encode($rawString): string
+	{
 		// Unpack string into an array of bytes
-		$bytes     = unpack( 'C*', $rawString );
-		$byteCount = count( $bytes );
+		$bytes     = unpack('C*', $rawString);
+		$byteCount = count($bytes);
 
 		$encodedString = '';
-		$byte          = array_shift( $bytes );
+		$byte          = array_shift($bytes);
 		$bitsRead      = 0;
 		$oldBits       = 0;
 
@@ -139,24 +141,24 @@ class Model_Base2n {
 		$encodedLength = $byteCount * $charsPerByte;
 
 		// Generate encoded output; each loop produces one encoded character
-		for ( $c = 0; $c < $encodedLength; $c++ ) {
+		for ($c = 0; $c < $encodedLength; $c++) {
 
 			// Get the bits needed for this encoded character
-			if ( $bitsRead + $bitsPerCharacter > 8 ) {
+			if ($bitsRead + $bitsPerCharacter > 8) {
 				// Not enough bits remain in this byte for the current character
 				// Save the remaining bits before getting the next byte
 				$oldBitCount = 8 - $bitsRead;
-				$oldBits     = $byte ^ ( $byte >> $oldBitCount << $oldBitCount );
+				$oldBits     = $byte ^ ($byte >> $oldBitCount << $oldBitCount);
 				$newBitCount = $bitsPerCharacter - $oldBitCount;
 
-				if ( ! $bytes ) {
+				if (! $bytes) {
 					// Last bits; match final character and exit loop
-					if ( $rightPadFinalBits ) {
+					if ($rightPadFinalBits) {
 						$oldBits <<= $newBitCount;
 					}
-					$encodedString .= $chars[ $oldBits ];
+					$encodedString .= $chars[$oldBits];
 
-					if ( $padFinalGroup ) {
+					if ($padFinalGroup) {
 						// Array of the lowest common multiples of $bitsPerCharacter and 8, divided by 8
 						$lcmMap         = array(
 							1 => 1,
@@ -168,16 +170,16 @@ class Model_Base2n {
 							7 => 7,
 							8 => 1,
 						);
-						$bytesPerGroup  = $lcmMap[ $bitsPerCharacter ];
-						$pads           = $bytesPerGroup * $charsPerByte - ceil( ( strlen( $rawString ) % $bytesPerGroup ) * $charsPerByte );
-						$encodedString .= str_repeat( $padCharacter, $pads );
+						$bytesPerGroup  = $lcmMap[$bitsPerCharacter];
+						$pads           = (int) ($bytesPerGroup * $charsPerByte - ceil((strlen($rawString) % $bytesPerGroup) * $charsPerByte));
+						$encodedString .= str_repeat($padCharacter, $pads);
 					}
 
 					break;
 				}
 
 				// Get next byte
-				$byte     = array_shift( $bytes );
+				$byte     = array_shift($bytes);
 				$bitsRead = 0;
 			} else {
 				$oldBitCount = 0;
@@ -185,16 +187,16 @@ class Model_Base2n {
 			}
 
 			// Read only the needed bits from this byte
-			$bits      = $byte >> 8 - ( $bitsRead + ( $newBitCount ) );
+			$bits      = $byte >> 8 - ($bitsRead + ($newBitCount));
 			$bits     ^= $bits >> $newBitCount << $newBitCount;
 			$bitsRead += $newBitCount;
 
-			if ( $oldBitCount ) {
+			if ($oldBitCount) {
 				// Bits come from seperate bytes, add $oldBits to $bits
-				$bits = ( $oldBits << $newBitCount ) | $bits;
+				$bits = ($oldBits << $newBitCount) | $bits;
 			}
 
-			$encodedString .= $chars[ $bits ];
+			$encodedString .= $chars[$bits];
 		}
 
 		return $encodedString;
@@ -207,8 +209,9 @@ class Model_Base2n {
 	 * @param   boolean $strict         Returns NULL if $encodedString contains an undecodable character
 	 * @return  string
 	 */
-	public function decode( $encodedString, $strict = false ): ?string {
-		if ( ! $encodedString || ! is_string( $encodedString ) ) {
+	public function decode($encodedString, $strict = false): ?string
+	{
+		if (! $encodedString || ! is_string($encodedString)) {
 			// Empty string, nothing to decode
 			return '';
 		}
@@ -222,25 +225,25 @@ class Model_Base2n {
 		$caseSensitive     = $this->_caseSensitive;
 
 		// Get index of encoded characters
-		if ( $this->_charmap ) {
+		if ($this->_charmap) {
 			$charmap = $this->_charmap;
 		} else {
 			$charmap = array();
 
-			for ( $i = 0; $i < $radix; $i++ ) {
-				$charmap[ $chars[ $i ] ] = $i;
+			for ($i = 0; $i < $radix; $i++) {
+				$charmap[$chars[$i]] = $i;
 			}
 
 			$this->_charmap = $charmap;
 		}
 
 		// The last encoded character is $encodedString[$lastNotatedIndex]
-		$lastNotatedIndex = strlen( $encodedString ) - 1;
+		$lastNotatedIndex = strlen($encodedString) - 1;
 
 		// Remove trailing padding characters
-		if ( $padFinalGroup ) {
-			while ( $encodedString[ $lastNotatedIndex ] === $padCharacter ) {
-				$encodedString = substr( $encodedString, 0, $lastNotatedIndex );
+		if ($padFinalGroup) {
+			while ($encodedString[$lastNotatedIndex] === $padCharacter) {
+				$encodedString = substr($encodedString, 0, $lastNotatedIndex);
 				--$lastNotatedIndex;
 			}
 		}
@@ -250,50 +253,50 @@ class Model_Base2n {
 		$bitsWritten = 0;
 
 		// Convert each encoded character to a series of unencoded bits
-		for ( $c = 0; $c <= $lastNotatedIndex; $c++ ) {
+		for ($c = 0; $c <= $lastNotatedIndex; $c++) {
 
-			if ( ! $caseSensitive && ! isset( $charmap[ $encodedString[ $c ] ] ) ) {
+			if (! $caseSensitive && ! isset($charmap[$encodedString[$c]])) {
 				// Encoded character was not found; try other case
-				if ( isset( $charmap[ $cUpper = strtoupper( $encodedString[ $c ] ) ] ) ) {
-					$charmap[ $encodedString[ $c ] ] = $charmap[ $cUpper ];
-				} elseif ( isset( $charmap[ $cLower = strtolower( $encodedString[ $c ] ) ] ) ) {
-					$charmap[ $encodedString[ $c ] ] = $charmap[ $cLower ];
+				if (isset($charmap[$cUpper = strtoupper($encodedString[$c])])) {
+					$charmap[$encodedString[$c]] = $charmap[$cUpper];
+				} elseif (isset($charmap[$cLower = strtolower($encodedString[$c])])) {
+					$charmap[$encodedString[$c]] = $charmap[$cLower];
 				}
 			}
 
-			if ( isset( $charmap[ $encodedString[ $c ] ] ) ) {
+			if (isset($charmap[$encodedString[$c]])) {
 				$bitsNeeded     = 8 - $bitsWritten;
 				$unusedBitCount = $bitsPerCharacter - $bitsNeeded;
 
 				// Get the new bits ready
-				if ( $bitsNeeded > $bitsPerCharacter ) {
+				if ($bitsNeeded > $bitsPerCharacter) {
 					// New bits aren't enough to complete a byte; shift them left into position
-					$newBits      = $charmap[ $encodedString[ $c ] ] << $bitsNeeded - $bitsPerCharacter;
+					$newBits      = $charmap[$encodedString[$c]] << $bitsNeeded - $bitsPerCharacter;
 					$bitsWritten += $bitsPerCharacter;
-				} elseif ( $c !== $lastNotatedIndex || $rightPadFinalBits ) {
+				} elseif ($c !== $lastNotatedIndex || $rightPadFinalBits) {
 					// Zero or more too many bits to complete a byte; shift right
-					$newBits     = $charmap[ $encodedString[ $c ] ] >> $unusedBitCount;
+					$newBits     = $charmap[$encodedString[$c]] >> $unusedBitCount;
 					$bitsWritten = 8; // $bitsWritten += $bitsNeeded;
 
 				} else {
 					// Final bits don't need to be shifted
-					$newBits     = $charmap[ $encodedString[ $c ] ];
+					$newBits     = $charmap[$encodedString[$c]];
 					$bitsWritten = 8;
 				}
 
 				$byte |= $newBits;
 
-				if ( $bitsWritten === 8 || $c === $lastNotatedIndex ) {
+				if ($bitsWritten === 8 || $c === $lastNotatedIndex) {
 					// Byte is ready to be written
-					$rawString .= pack( 'C', $byte );
+					$rawString .= pack('C', $byte);
 
-					if ( $c !== $lastNotatedIndex ) {
+					if ($c !== $lastNotatedIndex) {
 						// Start the next byte
 						$bitsWritten = $unusedBitCount;
-						$byte        = ( $charmap[ $encodedString[ $c ] ] ^ ( $newBits << $unusedBitCount ) ) << 8 - $bitsWritten;
+						$byte        = ($charmap[$encodedString[$c]] ^ ($newBits << $unusedBitCount)) << 8 - $bitsWritten;
 					}
 				}
-			} elseif ( $strict ) {
+			} elseif ($strict) {
 				// Unable to decode character; abort
 				return null;
 			}
