+++ classes/model/tokenbucket.php
@@ -2,7 +2,8 @@
 
 namespace TFAuthLS;
 
-class Model_TokenBucket {
+class Model_TokenBucket
+{
 
 	/* Constants to map from tokens per unit to tokens per second */
 	const MICROSECOND = 0.000001;
@@ -33,15 +34,16 @@ class Model_TokenBucket {
 	 * @param double $tokensPerSecond The number of tokens per second added to the bucket.
 	 * @param string $backing The backing storage to use.
 	 */
-	public function __construct( $identifier, $bucketSize, $tokensPerSecond, $backing = self::BACKING_WP_OPTIONS ) {
+	public function __construct($identifier, $bucketSize, $tokensPerSecond, $backing = self::BACKING_WP_OPTIONS)
+	{
 		$this->_identifier      = $identifier;
 		$this->_bucketSize      = $bucketSize;
 		$this->_tokensPerSecond = $tokensPerSecond;
 		$this->_backing         = $backing;
 
-		if ( $backing == self::BACKING_REDIS ) {
+		if ($backing == self::BACKING_REDIS) {
 			$this->_redis = new \Redis();
-			$this->_redis->pconnect( '127.0.0.1' );
+			$this->_redis->pconnect('127.0.0.1');
 		}
 	}
 
@@ -51,75 +53,79 @@ class Model_TokenBucket {
 	 * @param int $timeout
 	 * @return bool Whether or not the lock was acquired.
 	 */
-	private function _lock( $timeout = 30 ): bool {
-		if ( $this->_backing == self::BACKING_WP_OPTIONS ) {
-			$start = microtime( true );
-			while ( ! $this->_wp_options_create_lock( $this->_identifier ) ) {
-				if ( microtime( true ) - $start > $timeout ) {
+	private function _lock($timeout = 30): bool
+	{
+		if ($this->_backing == self::BACKING_WP_OPTIONS) {
+			$start = microtime(true);
+			while (! $this->_wp_options_create_lock($this->_identifier)) {
+				if (microtime(true) - $start > $timeout) {
 					return false;
 				}
-				usleep( 5000 ); // 5 ms
+				usleep(5000); // 5 ms
 			}
 			return true;
 		}
-		if ( $this->_backing == self::BACKING_REDIS ) {
-			if ( null === $this->_redis ) {
+		if ($this->_backing == self::BACKING_REDIS) {
+			if (null === $this->_redis) {
 				return false;
 			}
-			$start = microtime( true );
-			while ( ! $this->_redis->setnx( 'lock:' . $this->_identifier, '1' ) ) {
-				if ( microtime( true ) - $start > $timeout ) {
+			$start = microtime(true);
+			while (! $this->_redis->setnx('lock:' . $this->_identifier, '1')) {
+				if (microtime(true) - $start > $timeout) {
 					return false;
 				}
-				usleep( 5000 ); // 5 ms
+				usleep(5000); // 5 ms
 			}
-			$this->_redis->expire( 'lock:' . $this->_identifier, 30 );
+			$this->_redis->expire('lock:' . $this->_identifier, 30);
 			return true;
 		}
 		return false;
 	}
 
-	private function _unlock(): void {
-		if ( $this->_backing == self::BACKING_WP_OPTIONS ) {
-			$this->_wp_options_release_lock( $this->_identifier );
-		} elseif ( $this->_backing == self::BACKING_REDIS ) {
-			if ( null === $this->_redis ) {
+	private function _unlock(): void
+	{
+		if ($this->_backing == self::BACKING_WP_OPTIONS) {
+			$this->_wp_options_release_lock($this->_identifier);
+		} elseif ($this->_backing == self::BACKING_REDIS) {
+			if (null === $this->_redis) {
 				return;
 			}
-			$this->_redis->del( 'lock:' . $this->_identifier );
+			$this->_redis->del('lock:' . $this->_identifier);
 		}
 	}
 
-	private function _wp_options_create_lock( string $name, $timeout = null ) {
+	private function _wp_options_create_lock(string $name, $timeout = null)
+	{
 		// Our own version of WP_Upgrader::create_lock
 		global $wpdb;
 
-		if ( ! $timeout ) {
+		if (! $timeout) {
 			$timeout = 3600;
 		}
 
 		$lock_option = 'wfls_' . $name . '.lock';
-		$lock_result = $wpdb->query( $wpdb->prepare( "INSERT IGNORE INTO `{$wpdb->options}` (`option_name`, `option_value`, `autoload`) VALUES (%s, %s, 'no')", $lock_option, time() ) );
+		$lock_result = $wpdb->query($wpdb->prepare("INSERT IGNORE INTO `{$wpdb->options}` (`option_name`, `option_value`, `autoload`) VALUES (%s, %s, 'no')", $lock_option, time()));
 
-		if ( ! $lock_result ) {
-			$lock_result = get_option( $lock_option );
-			if ( ! $lock_result ) {
+		if (! $lock_result) {
+			$lock_result = get_option($lock_option);
+			if (! $lock_result) {
 				return false;
 			}
 
-			if ( $lock_result > ( time() - $timeout ) ) {
+			if ($lock_result > (time() - $timeout)) {
 				return false;
 			}
 
-			$this->_wp_options_release_lock( $name );
-			return $this->_wp_options_create_lock( $name, $timeout );
+			$this->_wp_options_release_lock($name);
+			return $this->_wp_options_create_lock($name, $timeout);
 		}
 
 		return true;
 	}
 
-	private function _wp_options_release_lock( string $name ) {
-		return delete_option( 'wfls_' . $name . '.lock' );
+	private function _wp_options_release_lock(string $name)
+	{
+		return delete_option('wfls_' . $name . '.lock');
 	}
 
 	/**
@@ -128,56 +134,58 @@ class Model_TokenBucket {
 	 * @param int $tokenCount
 	 * @return bool Whether or not there were enough tokens to satisfy the request.
 	 */
-	public function consume( $tokenCount = 1 ): bool {
-		if ( ! $this->_lock() ) {
+	public function consume($tokenCount = 1): bool
+	{
+		if (! $this->_lock()) {
 			return false;
 		}
 
-		if ( $this->_backing == self::BACKING_WP_OPTIONS ) {
-			$record = get_transient( 'wflsbucket:' . $this->_identifier );
-		} elseif ( $this->_backing == self::BACKING_REDIS ) {
-			$record = $this->_redis->get( 'bucket:' . $this->_identifier );
+		if ($this->_backing == self::BACKING_WP_OPTIONS) {
+			$record = get_transient('wflsbucket:' . $this->_identifier);
+		} elseif ($this->_backing == self::BACKING_REDIS) {
+			$record = $this->_redis->get('bucket:' . $this->_identifier);
 		} else {
 			$this->_unlock();
 			return false;
 		}
 
-		if ( $record === false ) {
-			if ( $tokenCount > $this->_bucketSize ) {
+		if ($record === false) {
+			if ($tokenCount > $this->_bucketSize) {
 				$this->_unlock();
 				return false;
 			}
 
-			$this->_bootstrap( $this->_bucketSize - $tokenCount );
+			$this->_bootstrap($this->_bucketSize - $tokenCount);
 			$this->_unlock();
 			return true;
 		}
 
-		$tokens = min( $this->_secondsToTokens( microtime( true ) - (float) $record ), $this->_bucketSize );
-		if ( $tokenCount > $tokens ) {
+		$tokens = min($this->_secondsToTokens(microtime(true) - (float) $record), $this->_bucketSize);
+		if ($tokenCount > $tokens) {
 			$this->_unlock();
 			return false;
 		}
 
-		if ( $this->_backing === self::BACKING_WP_OPTIONS ) {
-			set_transient( 'wflsbucket:' . $this->_identifier, (string) ( microtime( true ) - $this->_tokensToSeconds( $tokens - $tokenCount ) ), ceil( $this->_tokensToSeconds( $this->_bucketSize ) ) );
-		} elseif ( $this->_backing === self::BACKING_REDIS ) {
-			$this->_redis->set( 'bucket:' . $this->_identifier, (string) ( microtime( true ) - $this->_tokensToSeconds( $tokens - $tokenCount ) ) );
+		if ($this->_backing === self::BACKING_WP_OPTIONS) {
+			set_transient('wflsbucket:' . $this->_identifier, (string) (microtime(true) - $this->_tokensToSeconds($tokens - $tokenCount)), (int) ceil($this->_tokensToSeconds($this->_bucketSize)));
+		} elseif ($this->_backing === self::BACKING_REDIS) {
+			$this->_redis->set('bucket:' . $this->_identifier, (string) (microtime(true) - $this->_tokensToSeconds($tokens - $tokenCount)));
 		}
 
 		$this->_unlock();
 		return true;
 	}
 
-	public function reset(): ?bool {
-		if ( ! $this->_lock() ) {
+	public function reset(): ?bool
+	{
+		if (! $this->_lock()) {
 			return false;
 		}
 
-		if ( $this->_backing == self::BACKING_WP_OPTIONS ) {
-			delete_transient( 'wflsbucket:' . $this->_identifier );
-		} elseif ( $this->_backing == self::BACKING_REDIS ) {
-			$this->_redis->del( 'bucket:' . $this->_identifier );
+		if ($this->_backing == self::BACKING_WP_OPTIONS) {
+			delete_transient('wflsbucket:' . $this->_identifier);
+		} elseif ($this->_backing == self::BACKING_REDIS) {
+			$this->_redis->del('bucket:' . $this->_identifier);
 		}
 
 		$this->_unlock();
@@ -189,20 +197,23 @@ class Model_TokenBucket {
 	 *
 	 * @param int $initialTokens
 	 */
-	protected function _bootstrap( $initialTokens ) {
-		$microtime = microtime( true ) - $this->_tokensToSeconds( $initialTokens );
-		if ( $this->_backing == self::BACKING_WP_OPTIONS ) {
-			set_transient( 'wflsbucket:' . $this->_identifier, (string) $microtime, ceil( $this->_tokensToSeconds( $this->_bucketSize ) ) );
-		} elseif ( $this->_backing == self::BACKING_REDIS ) {
-			$this->_redis->set( 'bucket:' . $this->_identifier, (string) $microtime );
+	protected function _bootstrap($initialTokens)
+	{
+		$microtime = microtime(true) - $this->_tokensToSeconds($initialTokens);
+		if ($this->_backing == self::BACKING_WP_OPTIONS) {
+			set_transient('wflsbucket:' . $this->_identifier, (string) $microtime, (int) ceil($this->_tokensToSeconds($this->_bucketSize)));
+		} elseif ($this->_backing == self::BACKING_REDIS) {
+			$this->_redis->set('bucket:' . $this->_identifier, (string) $microtime);
 		}
 	}
 
-	protected function _tokensToSeconds( $tokens ): int|float {
+	protected function _tokensToSeconds($tokens): int|float
+	{
 		return $tokens / $this->_tokensPerSecond;
 	}
 
-	protected function _secondsToTokens( $seconds ): int|float {
+	protected function _secondsToTokens($seconds): int|float
+	{
 		return (int) $seconds * $this->_tokensPerSecond;
 	}
 }
